Add description

+++ src/Sylius/Bundle/CoreBundle/Migrations/Version20220203115813.php
@@ -14,14 +14,14 @@ final class Version20220203115813 extends AbstractMigration
 {
     public function getDescription(): string
     {
-        return '';
+        return 'Add messenger transport table for doctrine transports.';
     }
 
     public function up(Schema $schema): void
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on \'mysql\'.');
 
-        if (array_key_exists('MESSENGER_TRANSPORT_DSN', $_ENV) && str_contains($_ENV['MESSENGER_TRANSPORT_DSN'], 'doctrine')) {
+        if ($this->isUsingDoctrineTransport() && !$schema->hasTable('messenger_messages')) {
             $this->addSql('CREATE TABLE messenger_messages (id BIGINT AUTO_INCREMENT NOT NULL, body LONGTEXT NOT NULL, headers LONGTEXT NOT NULL, queue_name VARCHAR(255) NOT NULL, created_at DATETIME NOT NULL, available_at DATETIME NOT NULL, delivered_at DATETIME DEFAULT NULL, INDEX IDX_75EA56E016BA31DB (delivered_at), PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB');
         }
     }
@@ -30,6 +30,13 @@ final class Version20220203115813 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on \'mysql\'.');
 
-        $this->addSql('DROP TABLE messenger_messages');
+        if ($this->isUsingDoctrineTransport()) {
+            $this->addSql('DROP TABLE messenger_messages');
+        }
+    }
+
+    private function isUsingDoctrineTransport(): bool
+    {
+        return array_key_exists('MESSENGER_TRANSPORT_DSN', $_ENV) && str_contains($_ENV['MESSENGER_TRANSPORT_DSN'], 'doctrine');
     }
 }
